Fix DomainEvents tests against 1.6 until workaround is available
Fix DomainEvents tests against 1.6 until workaround is available

Fix DomainEvents tests against 1.6 until workaround is available
